+++ paradise-enhance/src/main/java/com/github/houbb/paradise/enhance/constant/translator/GoogleLanguageEnum.java
@@ -1,0 +1,137 @@
+package com.github.houbb.paradise.enhance.constant.translator;
+
+/**
+ * google 翻译语言枚举
+ * @author bbhou
+ * @version 1.1.1
+ */
+public enum GoogleLanguageEnum {
+    AUTO_DETECT(""),
+    AFRIKAANS("af"),
+    ALBANIAN("sq"),
+    AMHARIC("am"),
+    ARABIC("ar"),
+    ARMENIAN("hy"),
+    AZERBAIJANI("az"),
+    BASQUE("eu"),
+    BELARUSIAN("be"),
+    BENGALI("bn"),
+//    BIHARI("bh"),
+    BULGARIAN("bg"),
+    BURMESE("my"),
+    CATALAN("ca"),
+//    CHEROKEE("chr"),
+    CHINESE("zh"),
+    CHINESE_SIMPLIFIED("zh-CN"),
+    CHINESE_TRADITIONAL("zh-TW"),
+    CROATIAN("hr"),
+    CZECH("cs"),
+    DANISH("da"),
+//    DHIVEHI("dv"),
+    DUTCH("nl"),
+    ENGLISH("en"),
+    ESPERANTO("eo"),
+    ESTONIAN("et"),
+    FILIPINO("tl"),
+    FINNISH("fi"),
+    FRENCH("fr"),
+    GALICIAN("gl"),
+    GEORGIAN("ka"),
+    GERMAN("de"),
+    GREEK("el"),
+//    GUARANI("gn"),
+    GUJARATI("gu"),
+    HEBREW("iw"),
+    HINDI("hi"),
+    HUNGARIAN("hu"),
+    ICELANDIC("is"),
+    INDONESIAN("id"),
+//    INUKTITUT("iu"),
+    IRISH("ga"),
+    ITALIAN("it"),
+    JAPANESE("ja"),
+    KANNADA("kn"),
+    KAZAKH("kk"),
+    KHMER("km"),
+    KOREAN("ko"),
+    KURDISH("ku"),
+    KYRGYZ("ky"),
+    LAOTHIAN("lo"),
+    LATVIAN("lv"),
+    LITHUANIAN("lt"),
+    MACEDONIAN("mk"),
+    MALAY("ms"),
+    MALAYALAM("ml"),
+    MALTESE("mt"),
+    MARATHI("mr"),
+    MONGOLIAN("mn"),
+    NEPALI("ne"),
+    NORWEGIAN("no"),
+//    ORIYA("or"),
+    PASHTO("ps"),
+    PERSIAN("fa"),
+    POLISH("pl"),
+    PORTUGUESE("pt"),
+    PUNJABI("pa"),
+    ROMANIAN("ro"),
+    RUSSIAN("ru"),
+//    SANSKRIT("sa"),
+    SERBIAN("sr"),
+    SINDHI("sd"),
+    SINHALESE("si"),
+    SLOVAK("sk"),
+    SLOVENIAN("sl"),
+    SPANISH("es"),
+    SWAHILI("sw"),
+    SWEDISH("sv"),
+    TAJIK("tg"),
+    TAMIL("ta"),
+    TAGALOG("tl"),
+    TELUGU("te"),
+    THAI("th"),
+//    TIBETAN("bo"),
+    TURKISH("tr"),
+    UKRANIAN("uk"),
+    URDU("ur"),
+    UZBEK("uz"),
+//    UIGHUR("ug"),
+    VIETNAMESE("vi"),
+    WELSH("cy"),
+    YIDDISH("yi");
+
+    /**
+     * 语言类型
+     */
+    private String type;
+
+    GoogleLanguageEnum(String value)
+    {
+        this.type = value;
+    }
+
+    public static GoogleLanguageEnum fromString(final String pLanguage) {
+        for (GoogleLanguageEnum lang : values()) {
+            if (pLanguage.equals(lang.langInfo())) {
+                return lang ;
+            }
+        }
+        return null;
+    }
+
+    /**
+     * 显示语言信息
+     * @return
+     */
+    public String langInfo()
+    {
+        return this.type;
+    }
+
+    @Override
+    public String toString() {
+        return "GoogleLanguageEnum{" +
+                "type='" + type + '\'' +
+                '}';
+    }
+
+}