+++ paradise-enhance/src/main/java/com/github/houbb/paradise/enhance/core/connector/Connector.java
@@ -1,0 +1,7 @@
+package com.github.houbb.paradise.enhance.core.connector;
+
+public interface Connector<T> {
+
+    T connect();
+
+}