+++ paradise-enhance/src/main/java/com/github/houbb/paradise/enhance/core/splitter/Splitter.java
@@ -1,0 +1,9 @@
+package com.github.houbb.paradise.enhance.core.splitter;
+
+import java.util.List;
+
+public interface Splitter {
+
+    List<String> split();
+
+}