+++ paradise-enhance/src/main/java/com/github/houbb/paradise/enhance/core/translator/Translator.java
@@ -1,0 +1,7 @@
+package com.github.houbb.paradise.enhance.core.translator;
+
+public interface Translator<T> {
+
+    T translate() throws Exception;
+
+}