+++ paradise-enhance/src/main/java/com/github/houbb/paradise/enhance/core/translator/impl/YoudaoTranslator.java
@@ -1,0 +1,4 @@
+package com.github.houbb.paradise.enhance.core.translator.impl;
+
+public class YoudaoTranslator {
+}